pyproject.toml
[project]
name = "avito"
version = "0.1.0"
description = ""
authors = [
    {name = "frukun"}
]
readme = "README.md"
requires-python = "3.13.5"
dependencies = [
    "flask (>=3.1.1,<4.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "validators (>=0.35.0,<0.36.0)",
    "requests (>=2.32.4,<3.0.0)",
    "waitress (>=3.0.2,<4.0.0)",
    "flask-cors (>=6.0.1,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-xdist = "^3.7.0"
pytest-cov = "^6.2.1"
coverage-badge = "^1.1.2"
pytest-mock = "^3.14.1"


[tool.poetry.group.postgres.dependencies]
psycopg = {extras = ["binary"], version = "^3.2.9"}

[tool.pytest.ini_options]
addopts = ['-ra']
pythonpath = "."
